fix(test/tap): batch 2 of static analysis real-bug fixes
Batch 1 - dead stores/unused vars (11 files):
- remove dead test_count assignments in genai_module/async/embedding_rerank/mcp_module
- remove unused result strings in ai_llm_retry_scenarios
- remove dead jitter_ms in ai_llm_retry_scenarios
- remove unused row assignments in mysql-fast_forward/mysql-mirror1
- remove dead rc, sock, num_rows in mysql-protocol_compression/malformed/ssl_CA

Batch 2 - narrowing conversions (~100 fixes):
- pgsql-extended_query_protocol_test: uint8_t->char, ssize_t->int casts
- genai_module/async/embedding_rerank: recv()/mysql_num_rows() casts
- mysql-protocol_compression_level: various narrowing casts
- max_connections_ff: timeout/duration casts
- mysql_query_logging_memory: errno/plan() casts
- pgsql-connection_parameters: htonl/ntohl casts
- multiple_prepared_statements, mysql-reg_test_4716, nl2sql_internal, etc.

Batch 3 - macros + branch clone:
- fix macro parens in mysql-reg_test_4867_query_rules
- fix macro parens in proxysql_debug.h
- fix branch clone in pgsql-connection_parameters_test
- fix branch clone in pgsql-copy_to_test
- fix narrowing in pgsql-copy_from_test

Batch 4 - value params + widening + reserved IDs:
- pass string params by const ref in admin_various_commands2,
  mysql_hostgroup_attributes-servers_defaults,
  mysql-protocol_compression_level
- fix implicit widening overflow in mcp_mysql_concurrency_stress
- fix widening in mysql_stmt_send_long_data_large
- rename reserved identifiers (__1, _PtrSize_t)
- fix narrowing in pg_lite_client.h

+++ test/tap/tests/max_connections_ff-t.cpp
@@ -356,7 +356,7 @@ int test_ff_sess_exceeds_max_conns(const CommandLine& cl, MYSQL* proxy_admin, lo
 		connect_timeout_server, srv_conn_to, connect_timeout
 	);
 
-	my_err = set_srv_conn_to(proxy_admin, srv_conn_to);
+	my_err = set_srv_conn_to(proxy_admin, static_cast<int>(srv_conn_to));
 	if (my_err) {
 		diag("Failed to set 'mysql-connect_timeout_server' to '%ld'", srv_conn_to);
 		res = EXIT_FAILURE;
@@ -402,19 +402,19 @@ int test_ff_sess_exceeds_max_conns(const CommandLine& cl, MYSQL* proxy_admin, lo
 
 		diag("Executing query on the (n+1)th fast_forward connection...");
 		int q_err = mysql_query(proxy_ff, "DO 1");
-		int m_errno = mysql_errno(proxy_ff);
+		int m_errno = static_cast<int>(mysql_errno(proxy_ff));
 		const char* m_error = mysql_error(proxy_ff);
 
 		hrc::time_point end = hrc::now();
 
 		duration = end - start;
-		double duration_s = duration.count() / pow(10,9);
+		double duration_s = static_cast<int>(duration.count()) / pow(10,9);
 
 		diag("Query completed - Error: %d, ErrMsg: %s", m_errno, m_error);
 		diag("Time waited: %lf seconds", duration_s);
 
-		const double srv_conn_to_s = connect_timeout / 1000.0;
-		const double poll_to_s = poll_timeout / 1000.0;
+		const double srv_conn_to_s = static_cast<int>(connect_timeout) / 1000.0;
+		const double poll_to_s = static_cast<int>(poll_timeout) / 1000.0;
 		const double grace = 500 / 1000.0;
 
 		ok(
